+++ src/views/film/Index.vue
@@ -2,7 +2,7 @@
 import { ref } from 'vue'
 import { useRouter } from "vue-router"
 import { errorNotif } from '../../utils/notification'
-import { User } from '../../utils/user'
+import { User, isAdministrator } from '../../utils/user'
 
 const data = ref([])
 const router = useRouter()
@@ -21,6 +21,11 @@ function getData() {
 }
 
 function supprimer(id) {
+  if (!isAdministrator()) {
+    errorNotif("Vous n'avez pas les droits pour supprimer ce film.")
+    router.replace({ name: "filmIndex" })
+    return
+  }
   const optionRequest = User.generateHeaders()
   optionRequest.method = 'DELETE'
   console.log(id)
@@ -42,7 +47,7 @@ getData() // On charge les données au chargement de la page
 
 <template>
   <h1>Les Films</h1>
-  <RouterLink :to="{ name: 'filmCreate' }"><o-button variant="primary">➕ Ajouter un film</o-button></RouterLink>
+  <RouterLink v-if="isAdministrator()" :to="{ name: 'filmCreate' }"><o-button variant="primary">➕ Ajouter un film</o-button></RouterLink>
 
   <o-table :data="data">
 
@@ -77,11 +82,11 @@ getData() // On charge les données au chargement de la page
         Voir
       </RouterLink>
 
-      <RouterLink :to="{ name: 'filmEdit', params: { id: props.row._id } }" class="button is-warning">
+      <RouterLink v-if="isAdministrator()" :to="{ name: 'filmEdit', params: { id: props.row._id } }" class="button is-warning">
         Editer
       </RouterLink>
 
-      <div class="button is-danger" @click="supprimer(props.row._id)">Supprimer</div>
+      <div v-if="isAdministrator()" class="button is-danger" @click="supprimer(props.row._id)">Supprimer</div>
     </o-table-column>
 
   </o-table>
