+++ src/views/genre/Index.vue
@@ -2,7 +2,7 @@
 import { ref } from 'vue'
 import { useRouter } from "vue-router"
 import { errorNotif } from '../../utils/notification'
-import { User } from '../../utils/user'
+import { User, isAdministrator } from '../../utils/user'
 
 const data = ref([])
 const router = useRouter()
@@ -21,6 +21,11 @@ function getData() {
 }
 
 function supprimer(id) {
+  if (!isAdministrator()) {
+    errorNotif("Vous n'avez pas les droits pour supprimer cette séance.")
+    router.replace({ name: "genreIndex" })
+    return
+  }
   const optionRequest = User.generateHeaders()
   optionRequest.method = 'DELETE'
   console.log(id)
@@ -42,7 +47,7 @@ getData() // On charge les données au chargement de la page
 
 <template>
   <h1>Les Genres</h1>
-  <RouterLink :to="{ name: 'genreCreate' }"><o-button variant="primary">➕ Ajouter un genre</o-button></RouterLink>
+  <RouterLink v-if="isAdministrator()" :to="{ name: 'genreCreate' }"><o-button variant="primary">➕ Ajouter un genre</o-button></RouterLink>
 
   <o-table :data="data">
 
@@ -50,11 +55,11 @@ getData() // On charge les données au chargement de la page
       {{ props.row.title }}
     </o-table-column>
 
-    <o-table-column field="created_at" label="Créé le :" v-slot:default="props">
+    <o-table-column field="created_at" label="Créé le :" sortable v-slot:default="props">
       {{ new Date(props.row.created_at).toLocaleString() }}
     </o-table-column>
 
-    <o-table-column field="updated_at" label="Mise à jour le :" v-slot:default="props">
+    <o-table-column field="updated_at" label="Mise à jour le :" sortable v-slot:default="props">
       {{ new Date(props.row.updated_at).toLocaleString() }}
     </o-table-column>
 
@@ -63,11 +68,11 @@ getData() // On charge les données au chargement de la page
           Voir
         </RouterLink>
       
-      <RouterLink :to="{ name: 'genreEdit', params: { id: props.row._id } }" class="button is-warning">
+      <RouterLink v-if="isAdministrator()" :to="{ name: 'genreEdit', params: { id: props.row._id } }" class="button is-warning">
         Editer
       </RouterLink>
 
-      <div class="button is-danger" @click="supprimer(props.row._id)">Supprimer</div>
+      <div v-if="isAdministrator()" class="button is-danger" @click="supprimer(props.row._id)">Supprimer</div>
     </o-table-column>
 
   </o-table>
